// baby_tracker_fe/src/openapi/models/Feed.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    FeedTypes,
    FeedTypesFromJSON,
    FeedTypesFromJSONTyped,
    FeedTypesToJSON,
} from './';

/**
 * 
 * @export
 * @interface Feed
 */
export interface Feed {
    /**
     * 
     * @type {number}
     * @memberof Feed
     */
    id?: number;
    /**
     * 
     * @type {Date}
     * @memberof Feed
     */
    startAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Feed
     */
    endAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof Feed
     */
    amount?: number;
    /**
     * 
     * @type {FeedTypes}
     * @memberof Feed
     */
    type: FeedTypes;
    /**
     * 
     * @type {number}
     * @memberof Feed
     */
    babyId?: number;
}

export function FeedFromJSON(json: any): Feed {
    return FeedFromJSONTyped(json, false);
}

export function FeedFromJSONTyped(json: any, ignoreDiscriminator: boolean): Feed {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'startAt': !exists(json, 'start_at') ? undefined : (new Date(json['start_at'])),
        'endAt': !exists(json, 'end_at') ? undefined : (new Date(json['end_at'])),
        'amount': !exists(json, 'amount') ? undefined : json['amount'],
        'type': FeedTypesFromJSON(json['type']),
        'babyId': !exists(json, 'baby_id') ? undefined : json['baby_id'],
    };
}

export function FeedToJSON(value?: Feed | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'start_at': value.startAt === undefined ? undefined : (value.startAt.toISOString()),
        'end_at': value.endAt === undefined ? undefined : (value.endAt.toISOString()),
        'amount': value.amount,
        'type': FeedTypesToJSON(value.type),
        'baby_id': value.babyId,
    };
}